api/requirements.txt
starlite
keycloak-basic
requests
pymongo
fastapi-restful